[stdlib] Move JS boxed Long implementation into a separate package
Here's what we are going to do: we are going to have two packages.
- `kotlin.js.internal.boxedLong`
- `kotlin.js.internal.longAsBigInt`

Each of these packages will contain the same functions,
but implemented differently, depending on the Long representation.

The compiler will choose from which package to use a function,
depending on the configured JavaScript target (ES5 or ES6).

KT-57128